+++ NEWS.md
@@ -5,6 +5,10 @@
 * Let pipeline perform raw data parsing and indices calculation on combined raw data from single tasks ([#73](https://github.com/psychelzh/tarflow.iquizoo/issues/73)).
 * Let `prepare_fetch_data()` signal error when `contents` contains no data.
 
+## New Features
+
+* Support pre-fetched `contents` as contents template. Call `I()` to input literal `data.frame` in `contents` template. This is useful if you have already fetched the contents data and want to use it directly.
+
 # tarflow.iquizoo 3.4.0
 
 ## New Features